Maybe the cache?
Running into issues where the iframe doesn't really work unless it's a new browser. It sometimes does, but I'm pretty sure it's just a cache issue. I'll clear it in real-time later, but for now, it should work

I also added a random Python server to spawn as a quick and dirty solution if it happens mid usage